+++ src/features/favoriteSlice/favoriteSlice.jsx
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit'
 
 const initialState = {
-
+    favorite: true,
     favoriteItem: {},
     favoriteList: []
 
@@ -11,7 +11,9 @@ const favoriteSlice = createSlice({
     name: "favorite",
     initialState,
     reducers: {
-
+        setFavorite: (state) => {
+            state.favorite = !state.favorite
+        },
         setFavoriteList: (state, { payload }) => {
             state.favoriteList = payload
         },
